pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
pytest = "^6.1.1"
pytest-dotenv = "^0.5.2"
mock = "^4.0.3"
requests-mock = "^1.8.0"
beautifulsoup4 = "^4.9.3"
selenium = "^3.141.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
